Move PIP_VERSION to pip_compat module
To DRY
