+++ piptools/_compat/pip_compat.py
@@ -4,6 +4,8 @@ import importlib
 import pip
 from pip._vendor.packaging.version import parse as parse_version
 
+PIP_VERSION = tuple(map(int, parse_version(pip.__version__).base_version.split(".")))
+
 
 def do_import(module_path, subimport=None, old_path=None):
     old_path = old_path or module_path
@@ -53,7 +55,7 @@ Session = do_import("_vendor.requests.sessions", "Session")
 Resolver = do_import("legacy_resolve", "Resolver", old_path="resolve")
 
 # pip 18.1 has refactored InstallRequirement constructors use by pip-tools.
-if parse_version(pip.__version__) < parse_version("18.1"):
+if PIP_VERSION < (18, 1):
     install_req_from_line = InstallRequirement.from_line
     install_req_from_editable = InstallRequirement.from_editable
 else:
