+++ tests/test_cli_compile.py
@@ -10,10 +10,9 @@ from pytest import mark
 
 from .utils import invoke
 
-from piptools._compat.pip_compat import path_to_url
+from piptools._compat.pip_compat import PIP_VERSION, path_to_url
 from piptools.repositories import PyPIRepository
 from piptools.scripts.compile import cli
-from piptools.utils import PIP_VERSION
 
 TEST_DATA_PATH = os.path.join(os.path.split(__file__)[0], "test_data")
 MINIMAL_WHEELS_PATH = os.path.join(TEST_DATA_PATH, "minimal_wheels")
